Local DB now works
In the people controller, I added an example to the Database only if it doesn't exist and then display all Persons on the People/Details page

+++ Models/Person.cs
@@ -1,13 +1,15 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 
 namespace TaxWebApp.Models
 {
     public class Person
     {
-        public int Id { get; set; }
+        //ID for database
+        public long Id { get; set; }
 
         public string FirstName { get; set; }
